[GCU] Open more information for instruction
