+++ paddle/fluid/framework/new_executor/pir_adaptor/pir_adaptor_util.h
@@ -46,11 +46,13 @@ using ExecutionConfig = interpreter::ExecutionConfig;
 class IfInstruction;
 class WhileInstruction;
 class PyLayerInstruction;
+class CustomEngineInstruction;
 class ValueExecutionInfo {
  public:
   friend class IfInstruction;
   friend class WhileInstruction;
   friend class PyLayerInstruction;
+  friend class CustomEngineInstruction;
 
   explicit ValueExecutionInfo(Scope* scope) : scope_(scope) {}
