+++ paddle/fluid/pir/utils/pd_op_to_kernel_utils.h
@@ -13,15 +13,21 @@
 // limitations under the License.
 #pragma once
 
-#include "paddle/fluid/pir/utils/pd_op_to_kernel_utils.h"
 #include "paddle/phi/common/place.h"
 #include "paddle/pir/include/core/program.h"
 
 namespace paddle {
 namespace dialect {
 
-TEST_API std::unique_ptr<pir::Program> PdOpLowerToKernelPass(
-    pir::Program* prog, phi::Place place = phi::CPUPlace());
+pir::Type ConvertOpTypeToKernelType(pir::Type op_type);
 
+void ProcessBlock(
+    const phi::Place& place,
+    pir::Block* block,
+    pir::Block* new_block,
+    pir::IrContext* ctx,
+    std::unordered_map<pir::Operation*, pir::Operation*>* map_op_pair,
+    std::unordered_map<pir::Value, pir::Value>* map_value_pair,
+    bool for_if_block = false);
 }  // namespace dialect
 }  // namespace paddle
