+++ paddle/fluid/pybind/pybind.cc
@@ -257,6 +257,7 @@ DECLARE_FILE_SYMBOLS(aligned_allocator);
 DECLARE_FILE_SYMBOLS(pass_timing);
 DECLARE_FILE_SYMBOLS(op_compatible_info);
 DECLARE_FILE_SYMBOLS(sub_graph_detector);
+DECLARE_FILE_SYMBOLS(pd_op_to_kernel_utils);
 
 namespace paddle::pybind {
