+++ python/setup.py.in
@@ -1180,6 +1180,7 @@ headers = (
     list(find_files('*.h','@PADDLE_SOURCE_DIR@/paddle/fluid/pir/dialect/operator/utils/'))+
     list(find_files('*.h','@PADDLE_SOURCE_DIR@/paddle/fluid/pir/dialect/kernel/ir/'))+
     list(find_files('*.h','@PADDLE_SOURCE_DIR@/paddle/pir/include/core/'))+
+    list(find_files('pd_op_to_kernel_utils.h','@PADDLE_SOURCE_DIR@/paddle/fluid/pir/utils/'))+
     #custom_engine
     list(find_files('custom_engine_ext.h','@PADDLE_SOURCE_DIR@/paddle/fluid/custom_engine/'))+
     #new_executor headers
